+++ app/src/androidTest/java/woowacourse/movie/result/view/MovieResultActivityTest.kt
@@ -12,8 +12,8 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
-import woowacourse.movie.utils.MovieIntentConstant.KEY_MOVIE_ID
-import woowacourse.movie.utils.MovieIntentConstant.KEY_MOVIE_RESERVATION_COUNT
+import woowacourse.movie.util.MovieIntentConstant.KEY_MOVIE_ID
+import woowacourse.movie.util.MovieIntentConstant.KEY_MOVIE_RESERVATION_COUNT
 
 @RunWith(AndroidJUnit4::class)
 class MovieResultActivityTest {
