+++ app/src/main/java/woowacourse/movie/detail/view/MovieDetailActivity.kt
@@ -7,15 +7,15 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.detail.model.MovieReservationCount
 import woowacourse.movie.detail.presenter.MovieDetailPresenter
 import woowacourse.movie.detail.presenter.contract.MovieDetailContract
-import woowacourse.movie.main.model.Movie
+import woowacourse.movie.model.Movie
+import woowacourse.movie.model.MovieReservationCount
 import woowacourse.movie.result.view.MovieResultActivity
-import woowacourse.movie.utils.MovieIntentConstant.INVALID_VALUE_MOVIE_ID
-import woowacourse.movie.utils.MovieIntentConstant.KEY_MOVIE_ID
-import woowacourse.movie.utils.MovieIntentConstant.KEY_MOVIE_RESERVATION_COUNT
-import woowacourse.movie.utils.formatTimestamp
+import woowacourse.movie.util.MovieIntentConstant.INVALID_VALUE_MOVIE_ID
+import woowacourse.movie.util.MovieIntentConstant.KEY_MOVIE_ID
+import woowacourse.movie.util.MovieIntentConstant.KEY_MOVIE_RESERVATION_COUNT
+import woowacourse.movie.util.formatTimestamp
 
 class MovieDetailActivity : AppCompatActivity(), MovieDetailContract.View {
     private lateinit var detailImage: ImageView
