+++ app/src/main/java/woowacourse/movie/model/MovieReservationCount.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.detail.model
+package woowacourse.movie.model
 
 class MovieReservationCount(val count: Int = DEFAULT_COUNT) {
     init {
