+++ app/src/test/java/woowacourse/movie/model/MovieReservationCountTest.kt
@@ -1,8 +1,7 @@
-package woowacourse.movie.main.model
+package woowacourse.movie.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import woowacourse.movie.detail.model.MovieReservationCount
 
 class MovieReservationCountTest {
     @Test
